Support bidirectional conversion of keysyms
Enhance ksymtocode() and convert_code() to accept a “direction”
parameter which can be one of TO_UNICODE, TO_8BIT or TO_AUTO.  The
latter enables the old behavior of using prefer_unicode.

This prepares the codebase for another patch to make loadkeys convert
diacriticals (for the KD?KBDIACRUC ioctls present in newer kernels).

Signed-off-by: Michael Schutte <michi@uiae.at>