Enable UNUMBERs in compose definitions
As always, auto-conversion applies in case of XLATE mode (or no
KDSKBDIACRUC support).

Since 18eadfe…, loadkeys supports the composition of two Unicode keysyms
to another Unicode keysym.  Whereas the resulting char could already be
specified as U+…, diacr and base had to be given as 8-bit characters,
which is ugly with today’s Unicode support.  Since it isn’t realistic to
make keymaps capable of UTF-8, I was asked by a Debian developer to at
least make it possible to specify Unicode codepoints.  rvalue is
unsuitable at this point (because it does some trickery with KT_LATIN
vs. KT_LETTER which is not applicable here), I decided to add this to a
new expression just for compose.

Signed-off-by: Michael Schutte <michi@uiae.at>
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>