Don't open console on --bkeymap
Check that -b hasn't specified, similarly to -m, before loadkeys tries to open
the console. As for --mktable, --bkeymap doesn't need it. This allows to dump
binary keymaps in a larger range of situations where ownership of the console
device is not guarranteed e.g., when running from an Xterm.

Signed-off-by: Olivier Mehani <olivier.mehani@nicta.com.au>
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>