Replace index() and rindex() with strchr() and strrchr()
Hi Alexey,

Attached is a patch for kbd-1.15.2. It replaces index() and rindex() with
strchr() and strrchr().

The man page of rindex says:
4.3BSD; marked as LEGACY in POSIX.1-2001. POSIX.1-2008 removes the
specifications of index() and rindex(), recommending strchr(3) and
strrchr(3) instead.

I tried to compile kbd for a uClibc based embedded system. First I got
compile errors about missing prototypes for index() and rindex(), which were
fixed by adding an include of <strings.h> (not <string.h> which was already
included). However, this only lead to a link error on index and rindex. It
seems uClibc, depending on some #defines, can either have no definition for
(r)index, a prototype (extern) declaration or a macro definition
substituting (r)index by str(r)chr.

I figured it was simpler to just replace index() and rindex() with strchr()
and strrchr() in the source code than to figure out what #defines are needed
to make it work. Most likely it is also more future proof. If you agree,
please apply the attached patch.

Signed-off-by: Maarten ter Huurne <maarten@treewalker.org>
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>