Remove old and obsolete fi.map
fi.map does not match current or past keyboard engravings (for example
'ä' produces '}'), some distributions ship it as fi-old.map, there is no
corresponding X keymap, and in general it does not look like a generic
purpose map. Its history is a bit hazy but Linus presumes it might be
based on a custom keymap variant he was using a long time ago.

Acked-by: Linus Torvalds <torvalds@linux-foundation.org>