Link against libintl when needed
Since the kbd code uses gettext functionality, it should link with the
libintl library using the @INTLLIBS@ symbol provided by the gettext m4
macro. This is necessary on C libraries that don't provide a built-in
gettext functionality, and rely on the libintl library provided by
gettext, as is the case with the uClibc C library.

@INTLLIBS@ is empty when linking a separate library is not needed, as
is the case with the glibc C library.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>