Problems with codepage.1
Broken command synopsis syntax.  This may mean you're using a
construction in the command synopsis other than the standard
[ ] | { }, or it may mean you have running text in the command synopsis
section (the latter is not technically an error, but most cases of it
are impossible to translate into DocBook markup), or it may mean the
command syntax fails to match the description.

The incorrect sytntax used to express option alternation cannot be lifted
to XML.  In addition, use of <> around the filename argument violates
the normal convention and is unparseable.

See http://catb.org/~esr/doclifter/bugs.html for details.