setfont: Add option -d to double size of font.
On high pixel density systems, such as 4k displays, typical font sizes like
8x16 are illegible to most readers.  One solution would be to install and
use fonts that are designed for these displays, but currently these are not
widely available and perhaps they will never be available in all the same
varieties as the low-density fonts that have been used for a long time.

This patch introduces an alternative: it implements a -d option to setfont
that causes each pixel to be doubled along each axis (which actually
quadruples the size of the font).  It makes all of the traditional fonts
usable on high-DPI display.

Signed-off-by: Ben Pfaff <blp@cs.stanford.edu>