Add solar24x32 font
Hi Alexey,
I would like to submit a new console font to you.
I had to create it because I have a HIDPI display and I got tired of squinting.
The font is 24x32 pixels, perfect for anything that is 200+ dpi.
It is called solar24x32v1 because it was originally based on sun12x22.
If the Sun font is public domain then mine is too.
This is a hand-crafted font and it is version 1.
I am still improving it and I will produce a second version in a few months.

Signed-off-by: Aristo Grenelle <fontism@mail.com>
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>