Add Canadian Multilingual Standard keyboard map
Hi,

I am emailing because I saw you are the maintainer of the Linux kernel
keyboard layouts. I noticed there was no regular console keymap for the
Canadian Multilingual Standard [1] so I had a go at creating one myself. I
thought it might be useful to others and could possibly be included in the kbd
project.

I have based it on the ca-multi layout [2] in Debian by Christian Perrier,
with changes to use the AltGr key instead of the Alt key and to match with the
layout used in Gnome (screenshot attached). Christian's is close to the
standard layout but has many differences.

I've attached a copy of the file, and one with the more traditional
left-aligned layout in case this is preferred.

[1] <https://en.wikipedia.org/wiki/QWERTY#Canadian_Multilingual_Standard>
[2] <https://sources.debian.org/src/console-data/2:1.12-6/keymaps/i386/qwerty/ca-multi.kmap/>

With kind regards,
Anthony

Signed-off-by: Anthony Perkins <anthony@acperkins.com>
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>