Show only one error message if the terminal is lost
If the user has asked via system call for the terminal to be hung up
(TIOCVHANGUP) in the process of executing loadkeys, then subsequent
ioctls that loadkeys executes will fail with an EIO error [1][2].

Since any further ioctls will return an EIO error, dozens of error
messages will be generated.

If we get an EIO error then there is no need to try to continue since we
no longer have a valid descriptor.

[1] https://github.com/torvalds/linux/blob/44c026a73be8038f03dbdeef028b642880cf1511/drivers/tty/tty_io.c#L2783
[2] https://github.com/torvalds/linux/blob/44c026a73be8038f03dbdeef028b642880cf1511/drivers/tty/tty_ldisc.c#L223

Link: https://github.com/legionus/kbd/issues/92
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>