Created new 'fr-latin9' based layout as new 'mac-fr'
Copied 'fr-latin9' as the basis for the new 'mac-fr' layout. Then tweaked the existing layout to fall more in line with the layout shipped on contemporary Macs (mine being a 2020 Macbook Air), as well as some of the finer details of the layout as it exists on mac OS. There may still be a few more tweaks to make in the future, but this is a great start, and can certainly be used without frustration.

Meant to address issue #73
See here for reference: https://www.apple.com/shop/product/MK2A3FC/A/magic-keyboard-french

Signed-off-by: Gabriel Bauer <gabeb1277@gmail.com>