Get a font with a charsize of at least 32
The new KD_FONT_OP_GET_TALL uses the height as the size for each
character (the old KD_FONT_OP_GET always uses 32). This works for fonts
larger than 32x32. But if we get the old 8x16 font using the
KD_FONT_OP_GET_TALL for it, then vpitch becomes 16. When saving such a
font to a file in psf format, the vpitch information will be lost and
the font will turn into garbage.

psf1 has no height information at all. psf2 preserves the height, but if
it's less than 32 we can't tell the vpitch for that font from a normal
font that has 32.

In the future we may add a flag to psf2 to distinguish fonts with vpitch
< 32, but this would require carrying this information until the font is
loaded into the linux kernel. There is currently no place to store this
information.

For now, we will use the KD_FONT_OP_GET for fonts, and only if the
font has a height greater than 32 will we use KD_FONT_OP_GET_TALL. In
this case, we can understand that we need to use psf2 to save such a
font and then, when reading, we can understand which interface should be
used to load into the kernel.

Fixes: 287a3bae ("font: Leverage KD_FONT_OP_GET/SET_TALL font operations")
Link: https://github.com/legionus/kbd/issues/94
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>