loadkeys: Don't look for keymap in the current directory
Now loadkeys first tries to look for a keymap in the current directory.
This behavior is not intuitive and leads to problems. It appears that
loadskeys triggers all inodes in the current cwd. This might lead to
deadlocks during boot, because the system might not be ready to trigger
an autofs this early.

To solve this I exclude the current directory from the keymaps search
paths.

We still have ways to specify a different directory to search for
keymaps than the one specified on build time. The user can set the
LOADKEYS_KEYMAP_PATH environment variable to specify a different search
path.

Link: https://github.com/legionus/kbd/issues/101
Signed-off-by: Alexey Gladkov <gladkov.alexey@gmail.com>