xkbsupport: Add keymap generation from the xkb database
Xorg/Wayland uses its own keyboard configuration mechanism. Because of
this, the configuration of the text console and Xorg layouts may differ.
Not all combinations of Xorg layouts have a keymap in the kbd database.
It's inconvenient for users.

To solve this problem, it is possible to try convert the Xorg keymap
to the kernel keymap.

The XKB (X Keyboard Extension) database can be accessed with
libxkbcommon, which does not depend on the X11 libraries.

libxkbcommon is a keyboard keymap compiler and support library which
processes a reduced subset of keymaps as defined by the XKB
specification [1].

libxkbcommon does not distribute a keymap dataset itself, other than for
testing purposes. The most common dataset is xkeyboard-config, which is
used by all current distributions for their X11 XKB data [2].

There is a project that uses a similar approach to unify configurations.
This is console-setup [3].

I talked with Anton Zinoviev and he helped me figure out their idea of
switching multiple layouts with a single key combination:

  > how you solved the problem of switching more than two layouts with
  > one key combination ?

  The keys with ISO_Next_Group, for example the right Alt (keynumber
  100), are defined like this:

  keycode 100 = ShiftL_lock
  ShiftL keycode 100 = ShiftR_lock
  ShiftR keycode 100 = ShiftR_lock
  ShiftL ShiftR keycode 100 = ShiftL_lock

  So this key will work in the following way:

  no modifiers are active -> ShiftL is active
  ShiftL is active -> ShiftL and ShiftR are active
  ShiftL and ShiftR are active -> ShiftR is active
  ShiftR is active -> no modifiers are active

  The groups (the layouts) are placed in the following way:

  active modifiers  1 group    2 groups    3 groups    4 groups
  ----------------  -------    --------    --------    --------
  no modifier       group 1    group 1     group 1     group 1
  ShiftL            group 1    group 2     group 2     group 2
  ShiftR            group 1    group 2     group 3     group 4
  ShiftL+ShiftR     group 1    group 1     group 1     group 3

  In this way we have the following cycles:

  1 group: gr1->gr1->gr1->...
  2 groups: gr1->gr2->gr1->gr2->...
  3 groups: gr1->gr2->gr1->gr4->gr1->gr2->gr1->gr4->...
  4 groups: gr1->gr2->gr3->gr4->gr1->gr2->gr3->gr4->...

  --
  Anton Zinoviev

The difference with this implementation is that I am not trying to read
and parse the XKB base files. I analyze the resulting xkb keymap.

[1] https://xkbcommon.org/doc/current/index.html
[2] https://www.freedesktop.org/wiki/Software/XKeyboardConfig/
[3] https://salsa.debian.org/installer-team/console-setup.git

Signed-off-by: Alexey Gladkov <legion@kernel.org>