data/consolefonts: Add a 16x32 size variant for eurlatgr
This PR adds a new version of the `eurlatgr` font with a 32px height.
Characters are simply scaled by x2, with no further changes.
This matches the approach followed by `latarcyrheb-sun`, which shares the same typeface as this font.

This allows the font to be usable in HIDPI displays, such as 4K laptops.

I've also added a symbolic link from `eurlatgr.psfu` to `eurlatgr-16.psfu`, so there's no
breaking changes.

Signed-off-by: José Relvas <josemonsantorelvas@gmail.com>