setfont: Rewrite arguments parser
The setfont utility is very old and has never used getopt or
getopt_long. The utility parses options using the old-fashioned method.
So the utility has options that are completely incompatible with getopt.
Over time, parsing options has become a real mess.

To fix it somehow, it is possible to create a getopt-like function that
will help to parse options but with more clear rules.

Signed-off-by: Alexey Gladkov <legion@kernel.org>