tests: Better test tty existance
It is not enough to use tty to check the terminal. It is necessary to
check whether the terminal points to a real terminal and not a
pseudo-terminal.

To do this, we use kbdinfo to obtain the terminal mode as a terminal
check. This will allow us to repeat the terminal search algorithm used
by all utilities in the project.

Link: https://github.com/legionus/kbd/issues/143
Signed-off-by: Alexey Gladkov <legion@kernel.org>