Preserve only necessary metadata during install
"cp -a" tries to preserve all attributes if possible. This implies
preserving mode, any ACL, xattr permissions, and security context.

In some cases (such as fakeroot), it is not possible to preserve that
all attributes are preserved.

  cp: failed to preserve ownership for /tmp/.pkgmk/pkg/usr/share/kbd/keymaps/i386/qwertz/sr-latin.map.gz: Operation not supported

To avoid this, we will require that only the necessary attributes be
preserved. It is important for us to preserve the simlinks.

Signed-off-by: Alexey Gladkov <legion@kernel.org>