Fix NULL reference in option parser
If an option that needs an argument is given, but the argument is
actually missing, the optind will increase by 1 and return, and if
the option is the last option, a NULL reference occurs at
setfont.c:290. This bug is fixed by returning '!' instead of '?' to
main and process this condition separately.

Signed-off-by: Bohai Li <lbhlbhlbh2002@icloud.com>