setfont: Initialize data pointer when resetting font
When resetting the font to default, the `data` field in the
`console_font_op` structure must be explicitly set to NULL.

Previously, this field was left uninitialized. The kernel treats a
non-NULL `data` pointer as a pointer to a font name string. This caused
the ioctl to fail with -EFAULT (Bad address) because it attempted to
read from a garbage address.

Signed-off-by: Yutao Zhang <tommyzhang4707@outlook.com>